pyproject.toml
[project]
name = "study-buddy-ai"
version = "0.1.0"
description = ""
authors = [
    {name = "Vladislav",email = "ogloblinvlad6@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.12"
dependencies = [
    "fastapi (>=0.135.2,<0.136.0)",
    "uvicorn (>=0.42.0,<0.43.0)",
    "optimum[onnxruntime] (>=1.17.0,<2.0.0)",
    "transformers (>=4.40.0,<5.0.0)",
    "faiss-cpu (>=1.13.2,<2.0.0)",
    "pdfplumber (>=0.11.9,<0.12.0)",
    "python-multipart (>=0.0.22,<0.0.23)",
    "python-docx (>=1.2.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-asyncio = "^0.24"
httpx = "^0.27"
ruff = "^0.6"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I"]  # pycodestyle errors, pyflakes, isort

[tool.ruff.lint.per-file-ignores]
"backend/main.py" = ["E402"]           # logging.basicConfig must run before other imports
"backend/tests/conftest.py" = ["E402"] # patch.start() must run before backend imports
"backend/scripts/*.py" = ["E501"]      # CLI command strings are intentionally long

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["backend/tests"]
